% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_parameters.R
\name{find_single_root}
\alias{find_single_root}
\title{ML parameters of exponential and weibull models applied to trees}
\usage{
find_single_root(data, mutation_position, rkvector, jack = FALSE,
  pack = "rootsolve", verbose = TRUE)
}
\arguments{
\item{data}{list of dataframes, produced by splitting read.csv(.._for_LRT.csv)}

\item{mutation_position}{end, start or middle. Defines position of mutation on a branch}
}
\value{
A named numeric vector c(p_root, p_precision,lambda_weib_root, lambda_exp_root)
}
\description{
ML parameters of exponential and weibull models applied to trees
}

